+++ Core/Module/CharacterData/Template/ITemplateHandler.cs
@@ -39,6 +39,6 @@ namespace Core.Module.CharacterData.Template
         Location GetInitialStartPoint();
         float GetCollisionRadius();
         float GetCollisionHeight();
-        double GetBaseHpRegen(int level);
+        float GetBaseHpRegen(int level);
     }
 }
